Making the version type long. Also updating to commons-lang 2.5

+++ greenbridge-junit/src/test/java/com/googlecode/greenbridge/annotation/AMPEDM_128_3.java
@@ -36,7 +36,7 @@ public class AMPEDM_128_3 implements ScenarioRef {
     }
 
     @Override
-    public int version() {
+    public long version() {
         return 3;
     }
